Extract event quality analysis into dedicated module
Core refactoring:
- Create lookout/core/event_quality.py with classify_event_quality()
- Move 130+ line function from rain_events.py to dedicated module
- Reduce rain_events.py from 789 to 660 lines (-129 lines)
- Add re-export for backward compatibility via __all__

Rationale: Separates quality metrics from catalog CRUD operations for better
maintainability. Event quality logic is self-contained with no dependencies
on RainEventCatalog class, enabling cleaner module organization.
